Generate Quarto HTML / PDF Country Factsheets in Batch

This function renders the 'Country_Factsheet' Quarto template for a list of countries, generating a report file for each.

Usage

generate_report(type = "donor", name = NULL)

Arguments

type

either donor recipient or destination.

name

name of donor recipient or destination if NULL batch process them

top_n

Top number of report to batch process based on total amount

Value

links to genrated reports.

Examples

# name = c("France, Government of", "Germany, Government of")
# ftsAnalysis::generate_report(type = "donor", name )
# ftsAnalysis::generate_report(type = "donor", name = "Switzerland, Government of")
# linkdonor <- generate_report(type = "donor")
# dput(linkdonor)
# linkrecipient <- generate_report(type = "recipient")
# dput(linkrecipient )
# linkdestination <- generate_report(type = "destination")
# dput(linkdestination)

linkdonor <- c("[Profile Report for donor: Algeria, Government of](../docs/reports/ftsAnalysis-donor-algeria-government-of.html)", 
"[Profile Report for donor: Armenia, Government of](../docs/reports/ftsAnalysis-donor-armenia-government-of.html)", 
"[Profile Report for donor: Australia, Government of](../docs/reports/ftsAnalysis-donor-australia-government-of.html)", 
"[Profile Report for donor: Austria, Government of](../docs/reports/ftsAnalysis-donor-austria-government-of.html)", 
"[Profile Report for donor: Azerbaijan, Government of](../docs/reports/ftsAnalysis-donor-azerbaijan-government-of.html)", 
"[Profile Report for donor: Belgium, Government of](../docs/reports/ftsAnalysis-donor-belgium-government-of.html)", 
"[Profile Report for donor: Bhutan, Government of](../docs/reports/ftsAnalysis-donor-bhutan-government-of.html)", 
"[Profile Report for donor: Brazil, Government of](../docs/reports/ftsAnalysis-donor-brazil-government-of.html)", 
"[Profile Report for donor: Brunei Darussalam, Government of](../docs/reports/ftsAnalysis-donor-brunei-darussalam-government-of.html)", 
"[Profile Report for donor: Bulgaria, Government of](../docs/reports/ftsAnalysis-donor-bulgaria-government-of.html)", 
"[Profile Report for donor: Burkina Faso, Government of](../docs/reports/ftsAnalysis-donor-burkina-faso-government-of.html)", 
"[Profile Report for donor: Cambodia, Government of](../docs/reports/ftsAnalysis-donor-cambodia-government-of.html)", 
"[Profile Report for donor: Canada, Government of](../docs/reports/ftsAnalysis-donor-canada-government-of.html)", 
"[Profile Report for donor: Chad, Government of](../docs/reports/ftsAnalysis-donor-chad-government-of.html)", 
"[Profile Report for donor: Chile, Government of](../docs/reports/ftsAnalysis-donor-chile-government-of.html)", 
"[Profile Report for donor: China, Government of](../docs/reports/ftsAnalysis-donor-china-government-of.html)", 
"[Profile Report for donor: Colombia, Government of](../docs/reports/ftsAnalysis-donor-colombia-government-of.html)", 
"[Profile Report for donor: Croatia, Government of](../docs/reports/ftsAnalysis-donor-croatia-government-of.html)", 
"[Profile Report for donor: Cyprus, Government of](../docs/reports/ftsAnalysis-donor-cyprus-government-of.html)", 
"[Profile Report for donor: Czech Republic, Government of](../docs/reports/ftsAnalysis-donor-czech-republic-government-of.html)", 
"[Profile Report for donor: Danish International Development Agency](../docs/reports/ftsAnalysis-donor-danish-international-development-agency.html)", 
"[Profile Report for donor: Denmark, Government of](../docs/reports/ftsAnalysis-donor-denmark-government-of.html)", 
"[Profile Report for donor: Ecuador, Government of](../docs/reports/ftsAnalysis-donor-ecuador-government-of.html)", 
"[Profile Report for donor: Education Above All Foundation](../docs/reports/ftsAnalysis-donor-education-above-all-foundation.html)", 
"[Profile Report for donor: Egypt, Government of](../docs/reports/ftsAnalysis-donor-egypt-government-of.html)", 
"[Profile Report for donor: Estonia, Government of](../docs/reports/ftsAnalysis-donor-estonia-government-of.html)", 
"[Profile Report for donor: Finland, Government of](../docs/reports/ftsAnalysis-donor-finland-government-of.html)", 
"[Profile Report for donor: France, Government of](../docs/reports/ftsAnalysis-donor-france-government-of.html)", 
"[Profile Report for donor: Germany, Government of](../docs/reports/ftsAnalysis-donor-germany-government-of.html)", 
"[Profile Report for donor: Greece, Government of](../docs/reports/ftsAnalysis-donor-greece-government-of.html)", 
"[Profile Report for donor: Guyana, Government of](../docs/reports/ftsAnalysis-donor-guyana-government-of.html)", 
"[Profile Report for donor: Haiti, Government of](../docs/reports/ftsAnalysis-donor-haiti-government-of.html)", 
"[Profile Report for donor: Holy See, Government of](../docs/reports/ftsAnalysis-donor-holy-see-government-of.html)", 
"[Profile Report for donor: Honduras, Government of](../docs/reports/ftsAnalysis-donor-honduras-government-of.html)", 
"[Profile Report for donor: Iceland, Government of](../docs/reports/ftsAnalysis-donor-iceland-government-of.html)", 
"[Profile Report for donor: Indonesia, Government of](../docs/reports/ftsAnalysis-donor-indonesia-government-of.html)", 
"[Profile Report for donor: Ireland, Government of](../docs/reports/ftsAnalysis-donor-ireland-government-of.html)", 
"[Profile Report for donor: Italy, Government of](../docs/reports/ftsAnalysis-donor-italy-government-of.html)", 
"[Profile Report for donor: Japan, Government of](../docs/reports/ftsAnalysis-donor-japan-government-of.html)", 
"[Profile Report for donor: Jersey Overseas Aid](../docs/reports/ftsAnalysis-donor-jersey-overseas-aid.html)", 
"[Profile Report for donor: Jordan, Government of](../docs/reports/ftsAnalysis-donor-jordan-government-of.html)", 
"[Profile Report for donor: Korea, Republic of, Government of](../docs/reports/ftsAnalysis-donor-korea-republic-of-government-of.html)", 
"[Profile Report for donor: Kuwait Fund for Arab Economic Development](../docs/reports/ftsAnalysis-donor-kuwait-fund-for-arab-economic-development.html)", 
"[Profile Report for donor: Kuwait, Government of](../docs/reports/ftsAnalysis-donor-kuwait-government-of.html)", 
"[Profile Report for donor: Kyrgyzstan, Government of](../docs/reports/ftsAnalysis-donor-kyrgyzstan-government-of.html)", 
"[Profile Report for donor: Latvia, Government of](../docs/reports/ftsAnalysis-donor-latvia-government-of.html)", 
"[Profile Report for donor: Liechtenstein, Government of](../docs/reports/ftsAnalysis-donor-liechtenstein-government-of.html)", 
"[Profile Report for donor: Lithuania, Government of](../docs/reports/ftsAnalysis-donor-lithuania-government-of.html)", 
"[Profile Report for donor: Luxembourg, Government of](../docs/reports/ftsAnalysis-donor-luxembourg-government-of.html)", 
"[Profile Report for donor: Madagascar, Government of](../docs/reports/ftsAnalysis-donor-madagascar-government-of.html)", 
"[Profile Report for donor: Malawi, Government of](../docs/reports/ftsAnalysis-donor-malawi-government-of.html)", 
"[Profile Report for donor: Malta, Government of](../docs/reports/ftsAnalysis-donor-malta-government-of.html)", 
"[Profile Report for donor: Mauritius, Government of](../docs/reports/ftsAnalysis-donor-mauritius-government-of.html)", 
"[Profile Report for donor: Monaco, Government of](../docs/reports/ftsAnalysis-donor-monaco-government-of.html)", 
"[Profile Report for donor: Mongolia, Government of](../docs/reports/ftsAnalysis-donor-mongolia-government-of.html)", 
"[Profile Report for donor: Montenegro, Government of](../docs/reports/ftsAnalysis-donor-montenegro-government-of.html)", 
"[Profile Report for donor: Morocco, Government of](../docs/reports/ftsAnalysis-donor-morocco-government-of.html)", 
"[Profile Report for donor: Mozambique, Government of](../docs/reports/ftsAnalysis-donor-mozambique-government-of.html)", 
"[Profile Report for donor: Nepal, Government of](../docs/reports/ftsAnalysis-donor-nepal-government-of.html)", 
"[Profile Report for donor: Netherlands, Government of](../docs/reports/ftsAnalysis-donor-netherlands-government-of.html)", 
"[Profile Report for donor: New Zealand, Government of](../docs/reports/ftsAnalysis-donor-new-zealand-government-of.html)", 
"[Profile Report for donor: Nigeria, Government of](../docs/reports/ftsAnalysis-donor-nigeria-government-of.html)", 
"[Profile Report for donor: Norway, Government of](../docs/reports/ftsAnalysis-donor-norway-government-of.html)", 
"[Profile Report for donor: Oman, Government of](../docs/reports/ftsAnalysis-donor-oman-government-of.html)", 
"[Profile Report for donor: Palestinian territory, occupied](../docs/reports/ftsAnalysis-donor-palestinian-territory-occupied.html)", 
"[Profile Report for donor: Peru, Government of](../docs/reports/ftsAnalysis-donor-peru-government-of.html)", 
"[Profile Report for donor: Philippines, Government of](../docs/reports/ftsAnalysis-donor-philippines-government-of.html)", 
"[Profile Report for donor: Poland, Government of](../docs/reports/ftsAnalysis-donor-poland-government-of.html)", 
"[Profile Report for donor: Portugal, Government of](../docs/reports/ftsAnalysis-donor-portugal-government-of.html)", 
"[Profile Report for donor: Qatar Airways](../docs/reports/ftsAnalysis-donor-qatar-airways.html)", 
"[Profile Report for donor: Qatar Fund for Development](../docs/reports/ftsAnalysis-donor-qatar-fund-for-development.html)", 
"[Profile Report for donor: Qatar, Government of](../docs/reports/ftsAnalysis-donor-qatar-government-of.html)", 
"[Profile Report for donor: Russian Federation, Government of](../docs/reports/ftsAnalysis-donor-russian-federation-government-of.html)", 
"[Profile Report for donor: San Marino, Government of](../docs/reports/ftsAnalysis-donor-san-marino-government-of.html)", 
"[Profile Report for donor: Saudi Arabia (Kingdom of), Government of](../docs/reports/ftsAnalysis-donor-saudi-arabia-kingdom-of-government-of.html)", 
"[Profile Report for donor: Singapore, Government of](../docs/reports/ftsAnalysis-donor-singapore-government-of.html)", 
"[Profile Report for donor: Slovenia, Government of](../docs/reports/ftsAnalysis-donor-slovenia-government-of.html)", 
"[Profile Report for donor: Somalia, Government of](../docs/reports/ftsAnalysis-donor-somalia-government-of.html)", 
"[Profile Report for donor: Spain, Government of](../docs/reports/ftsAnalysis-donor-spain-government-of.html)", 
"[Profile Report for donor: Sudan, Government of](../docs/reports/ftsAnalysis-donor-sudan-government-of.html)", 
"[Profile Report for donor: Sweden, Government of](../docs/reports/ftsAnalysis-donor-sweden-government-of.html)", 
"[Profile Report for donor: Switzerland, Government of](../docs/reports/ftsAnalysis-donor-switzerland-government-of.html)", 
"[Profile Report for donor: Tanzania, Government of](../docs/reports/ftsAnalysis-donor-tanzania-government-of.html)", 
"[Profile Report for donor: Thailand, Government of](../docs/reports/ftsAnalysis-donor-thailand-government-of.html)", 
"[Profile Report for donor: Türkiye, Government of](../docs/reports/ftsAnalysis-donor-turkiye-government-of.html)", 
"[Profile Report for donor: Uganda, Government of](../docs/reports/ftsAnalysis-donor-uganda-government-of.html)", 
"[Profile Report for donor: Ukraine, Government of](../docs/reports/ftsAnalysis-donor-ukraine-government-of.html)", 
"[Profile Report for donor: United Arab Emirates, Government of](../docs/reports/ftsAnalysis-donor-united-arab-emirates-government-of.html)", 
"[Profile Report for donor: United Kingdom, Government of](../docs/reports/ftsAnalysis-donor-united-kingdom-government-of.html)", 
"[Profile Report for donor: United States Department of State](../docs/reports/ftsAnalysis-donor-united-states-department-of-state.html)", 
"[Profile Report for donor: United States of America, Government of](../docs/reports/ftsAnalysis-donor-united-states-of-america-government-of.html)", 
"[Profile Report for donor: Viet Nam, Government of](../docs/reports/ftsAnalysis-donor-viet-nam-government-of.html)", 
"[Profile Report for donor: Zimbabwe, Government of](../docs/reports/ftsAnalysis-donor-zimbabwe-government-of.html)"
)

linkrecipient <- c("[Profile Report for recipient: Food and Agriculture Organization of the United Nations](../docs/reports/ftsAnalysis-recipient-food-and-agriculture-organization-of-the-united-nations.html)", 
"[Profile Report for recipient: International Labour Organization](../docs/reports/ftsAnalysis-recipient-international-labour-organization.html)", 
"[Profile Report for recipient: International Organization for Migration](../docs/reports/ftsAnalysis-recipient-international-organization-for-migration.html)", 
"[Profile Report for recipient: International Trade Center](../docs/reports/ftsAnalysis-recipient-international-trade-center.html)", 
"[Profile Report for recipient: Multi-donor flexible humanitarian contribution (UNICEF Global Humanitarian Thematic)](../docs/reports/ftsAnalysis-recipient-multi-donor-flexible-humanitarian-contribution-unicef-global-humanitarian-thematic.html)", 
"[Profile Report for recipient: Multi-donor flexible humanitarian contribution (UNICEF Regional Humanitarian Thematic)](../docs/reports/ftsAnalysis-recipient-multi-donor-flexible-humanitarian-contribution-unicef-regional-humanitarian-thematic.html)", 
"[Profile Report for recipient: Office for the Coordination of Humanitarian Affairs](../docs/reports/ftsAnalysis-recipient-office-for-the-coordination-of-humanitarian-affairs.html)", 
"[Profile Report for recipient: Office of the High Commissioner for Human Rights](../docs/reports/ftsAnalysis-recipient-office-of-the-high-commissioner-for-human-rights.html)", 
"[Profile Report for recipient: Pan-American Health Organization (World Health Organization)](../docs/reports/ftsAnalysis-recipient-pan-american-health-organization-world-health-organization.html)", 
"[Profile Report for recipient: Specially Designated Contributions (OCHA)](../docs/reports/ftsAnalysis-recipient-specially-designated-contributions-ocha.html)", 
"[Profile Report for recipient: UN Action Against Sexual Violence in Conflict](../docs/reports/ftsAnalysis-recipient-un-action-against-sexual-violence-in-conflict.html)", 
"[Profile Report for recipient: UN Agencies (Confidential)](../docs/reports/ftsAnalysis-recipient-un-agencies-confidential.html)", 
"[Profile Report for recipient: UN Office for Disaster Risk Reduction](../docs/reports/ftsAnalysis-recipient-un-office-for-disaster-risk-reduction.html)", 
"[Profile Report for recipient: UN Women](../docs/reports/ftsAnalysis-recipient-un-women.html)", 
"[Profile Report for recipient: UNDP for UNDSS](../docs/reports/ftsAnalysis-recipient-undp-for-undss.html)", 
"[Profile Report for recipient: United Nations](../docs/reports/ftsAnalysis-recipient-united-nations.html)", 
"[Profile Report for recipient: United Nations Capital Development Fund](../docs/reports/ftsAnalysis-recipient-united-nations-capital-development-fund.html)", 
"[Profile Report for recipient: United Nations Children and Armed Conflict (Office of the Special Representative)](../docs/reports/ftsAnalysis-recipient-united-nations-children-and-armed-conflict-office-of-the-special-representative.html)", 
"[Profile Report for recipient: United Nations Children's Fund](../docs/reports/ftsAnalysis-recipient-united-nations-childrens-fund.html)", 
"[Profile Report for recipient: United Nations Department of Safety and Security](../docs/reports/ftsAnalysis-recipient-united-nations-department-of-safety-and-security.html)", 
"[Profile Report for recipient: United Nations Development Programme](../docs/reports/ftsAnalysis-recipient-united-nations-development-programme.html)", 
"[Profile Report for recipient: United Nations Disaster Assessment and Coordination](../docs/reports/ftsAnalysis-recipient-united-nations-disaster-assessment-and-coordination.html)", 
"[Profile Report for recipient: United Nations Economic and Social Commission for Asia and the Pacific](../docs/reports/ftsAnalysis-recipient-united-nations-economic-and-social-commission-for-asia-and-the-pacific.html)", 
"[Profile Report for recipient: United Nations Educational, Scientific and Cultural Organization](../docs/reports/ftsAnalysis-recipient-united-nations-educational-scientific-and-cultural-organization.html)", 
"[Profile Report for recipient: United Nations Environment Programme](../docs/reports/ftsAnalysis-recipient-united-nations-environment-programme.html)", 
"[Profile Report for recipient: United Nations High Commissioner for Refugees](../docs/reports/ftsAnalysis-recipient-united-nations-high-commissioner-for-refugees.html)", 
"[Profile Report for recipient: United Nations Human Settlements Programme (UN-HABITAT)](../docs/reports/ftsAnalysis-recipient-united-nations-human-settlements-programme-un-habitat.html)", 
"[Profile Report for recipient: United Nations Humanitarian Response Depot](../docs/reports/ftsAnalysis-recipient-united-nations-humanitarian-response-depot.html)", 
"[Profile Report for recipient: United Nations Industrial Development Organization](../docs/reports/ftsAnalysis-recipient-united-nations-industrial-development-organization.html)", 
"[Profile Report for recipient: United Nations Institute for Disarmament Research](../docs/reports/ftsAnalysis-recipient-united-nations-institute-for-disarmament-research.html)", 
"[Profile Report for recipient: United Nations Institute for Training and Research](../docs/reports/ftsAnalysis-recipient-united-nations-institute-for-training-and-research.html)", 
"[Profile Report for recipient: United Nations Joint Programme on HIV/AIDS](../docs/reports/ftsAnalysis-recipient-united-nations-joint-programme-on-hivaids.html)", 
"[Profile Report for recipient: United Nations Mine Action Service](../docs/reports/ftsAnalysis-recipient-united-nations-mine-action-service.html)", 
"[Profile Report for recipient: United Nations Office for Project Services](../docs/reports/ftsAnalysis-recipient-united-nations-office-for-project-services.html)", 
"[Profile Report for recipient: United Nations Office on Drugs and Crime](../docs/reports/ftsAnalysis-recipient-united-nations-office-on-drugs-and-crime.html)", 
"[Profile Report for recipient: United Nations Population Fund](../docs/reports/ftsAnalysis-recipient-united-nations-population-fund.html)", 
"[Profile Report for recipient: United Nations Relief and Works Agency for Palestine Refugees in the Near East](../docs/reports/ftsAnalysis-recipient-united-nations-relief-and-works-agency-for-palestine-refugees-in-the-near-east.html)", 
"[Profile Report for recipient: United Nations Voluntary Fund for Victims of Torture](../docs/reports/ftsAnalysis-recipient-united-nations-voluntary-fund-for-victims-of-torture.html)", 
"[Profile Report for recipient: World Food Programme](../docs/reports/ftsAnalysis-recipient-world-food-programme.html)", 
"[Profile Report for recipient: World Health Organization](../docs/reports/ftsAnalysis-recipient-world-health-organization.html)"
)

linkdestination <- c("[Profile Report for destination: Afghanistan](../docs/reports/ftsAnalysis-destination-afghanistan.html)", 
"[Profile Report for destination: Albania](../docs/reports/ftsAnalysis-destination-albania.html)", 
"[Profile Report for destination: Algeria](../docs/reports/ftsAnalysis-destination-algeria.html)", 
"[Profile Report for destination: Angola](../docs/reports/ftsAnalysis-destination-angola.html)", 
"[Profile Report for destination: Anguilla (United Kingdom)](../docs/reports/ftsAnalysis-destination-anguilla-united-kingdom.html)", 
"[Profile Report for destination: Argentina](../docs/reports/ftsAnalysis-destination-argentina.html)", 
"[Profile Report for destination: Armenia](../docs/reports/ftsAnalysis-destination-armenia.html)", 
"[Profile Report for destination: Aruba (Netherlands)](../docs/reports/ftsAnalysis-destination-aruba-netherlands.html)", 
"[Profile Report for destination: Australia](../docs/reports/ftsAnalysis-destination-australia.html)", 
"[Profile Report for destination: Azerbaijan](../docs/reports/ftsAnalysis-destination-azerbaijan.html)", 
"[Profile Report for destination: Bangladesh](../docs/reports/ftsAnalysis-destination-bangladesh.html)", 
"[Profile Report for destination: Belize](../docs/reports/ftsAnalysis-destination-belize.html)", 
"[Profile Report for destination: Benin](../docs/reports/ftsAnalysis-destination-benin.html)", 
"[Profile Report for destination: Bolivia, Plurinational State of](../docs/reports/ftsAnalysis-destination-bolivia-plurinational-state-of.html)", 
"[Profile Report for destination: Bosnia and Herzegovina](../docs/reports/ftsAnalysis-destination-bosnia-and-herzegovina.html)", 
"[Profile Report for destination: Botswana](../docs/reports/ftsAnalysis-destination-botswana.html)", 
"[Profile Report for destination: Brazil](../docs/reports/ftsAnalysis-destination-brazil.html)", 
"[Profile Report for destination: Brunei Darussalam](../docs/reports/ftsAnalysis-destination-brunei-darussalam.html)", 
"[Profile Report for destination: Bulgaria](../docs/reports/ftsAnalysis-destination-bulgaria.html)", 
"[Profile Report for destination: Burkina Faso](../docs/reports/ftsAnalysis-destination-burkina-faso.html)", 
"[Profile Report for destination: Burundi](../docs/reports/ftsAnalysis-destination-burundi.html)", 
"[Profile Report for destination: Cambodia](../docs/reports/ftsAnalysis-destination-cambodia.html)", 
"[Profile Report for destination: Cameroon](../docs/reports/ftsAnalysis-destination-cameroon.html)", 
"[Profile Report for destination: Cape Verde](../docs/reports/ftsAnalysis-destination-cape-verde.html)", 
"[Profile Report for destination: Central African Republic](../docs/reports/ftsAnalysis-destination-central-african-republic.html)", 
"[Profile Report for destination: Chad](../docs/reports/ftsAnalysis-destination-chad.html)", 
"[Profile Report for destination: Chile](../docs/reports/ftsAnalysis-destination-chile.html)", 
"[Profile Report for destination: China](../docs/reports/ftsAnalysis-destination-china.html)", 
"[Profile Report for destination: Colombia](../docs/reports/ftsAnalysis-destination-colombia.html)", 
"[Profile Report for destination: Comoros](../docs/reports/ftsAnalysis-destination-comoros.html)", 
"[Profile Report for destination: Congo](../docs/reports/ftsAnalysis-destination-congo.html)", 
"[Profile Report for destination: Cook Islands](../docs/reports/ftsAnalysis-destination-cook-islands.html)", 
"[Profile Report for destination: Costa Rica](../docs/reports/ftsAnalysis-destination-costa-rica.html)", 
"[Profile Report for destination: Côte d'Ivoire](../docs/reports/ftsAnalysis-destination-cote-divoire.html)", 
"[Profile Report for destination: Cuba](../docs/reports/ftsAnalysis-destination-cuba.html)", 
"[Profile Report for destination: Czech Republic](../docs/reports/ftsAnalysis-destination-czech-republic.html)", 
"[Profile Report for destination: Democratic Republic of the Congo](../docs/reports/ftsAnalysis-destination-democratic-republic-of-the-congo.html)", 
"[Profile Report for destination: Denmark](../docs/reports/ftsAnalysis-destination-denmark.html)", 
"[Profile Report for destination: Djibouti](../docs/reports/ftsAnalysis-destination-djibouti.html)", 
"[Profile Report for destination: Dominican Republic](../docs/reports/ftsAnalysis-destination-dominican-republic.html)", 
"[Profile Report for destination: Ecuador](../docs/reports/ftsAnalysis-destination-ecuador.html)", 
"[Profile Report for destination: Egypt](../docs/reports/ftsAnalysis-destination-egypt.html)", 
"[Profile Report for destination: El Salvador](../docs/reports/ftsAnalysis-destination-el-salvador.html)", 
"[Profile Report for destination: Equatorial Guinea](../docs/reports/ftsAnalysis-destination-equatorial-guinea.html)", 
"[Profile Report for destination: Eritrea](../docs/reports/ftsAnalysis-destination-eritrea.html)", 
"[Profile Report for destination: Eswatini](../docs/reports/ftsAnalysis-destination-eswatini.html)", 
"[Profile Report for destination: Ethiopia](../docs/reports/ftsAnalysis-destination-ethiopia.html)", 
"[Profile Report for destination: Fiji](../docs/reports/ftsAnalysis-destination-fiji.html)", 
"[Profile Report for destination: France](../docs/reports/ftsAnalysis-destination-france.html)", 
"[Profile Report for destination: Gabon](../docs/reports/ftsAnalysis-destination-gabon.html)", 
"[Profile Report for destination: Gambia](../docs/reports/ftsAnalysis-destination-gambia.html)", 
"[Profile Report for destination: Georgia](../docs/reports/ftsAnalysis-destination-georgia.html)", 
"[Profile Report for destination: Ghana](../docs/reports/ftsAnalysis-destination-ghana.html)", 
"[Profile Report for destination: Global](../docs/reports/ftsAnalysis-destination-global.html)", 
"[Profile Report for destination: Greece](../docs/reports/ftsAnalysis-destination-greece.html)", 
"[Profile Report for destination: Grenada](../docs/reports/ftsAnalysis-destination-grenada.html)", 
"[Profile Report for destination: Guatemala](../docs/reports/ftsAnalysis-destination-guatemala.html)", 
"[Profile Report for destination: Guinea](../docs/reports/ftsAnalysis-destination-guinea.html)", 
"[Profile Report for destination: Guinea-Bissau](../docs/reports/ftsAnalysis-destination-guinea-bissau.html)", 
"[Profile Report for destination: Guyana](../docs/reports/ftsAnalysis-destination-guyana.html)", 
"[Profile Report for destination: Haiti](../docs/reports/ftsAnalysis-destination-haiti.html)", 
"[Profile Report for destination: Honduras](../docs/reports/ftsAnalysis-destination-honduras.html)", 
"[Profile Report for destination: Hungary](../docs/reports/ftsAnalysis-destination-hungary.html)", 
"[Profile Report for destination: India](../docs/reports/ftsAnalysis-destination-india.html)", 
"[Profile Report for destination: Indonesia](../docs/reports/ftsAnalysis-destination-indonesia.html)", 
"[Profile Report for destination: Iran, Islamic Republic of](../docs/reports/ftsAnalysis-destination-iran-islamic-republic-of.html)", 
"[Profile Report for destination: Iraq](../docs/reports/ftsAnalysis-destination-iraq.html)", 
"[Profile Report for destination: Israel](../docs/reports/ftsAnalysis-destination-israel.html)", 
"[Profile Report for destination: Italy](../docs/reports/ftsAnalysis-destination-italy.html)", 
"[Profile Report for destination: Jamaica](../docs/reports/ftsAnalysis-destination-jamaica.html)", 
"[Profile Report for destination: Jordan](../docs/reports/ftsAnalysis-destination-jordan.html)", 
"[Profile Report for destination: Kazakhstan](../docs/reports/ftsAnalysis-destination-kazakhstan.html)", 
"[Profile Report for destination: Kenya](../docs/reports/ftsAnalysis-destination-kenya.html)", 
"[Profile Report for destination: Kiribati](../docs/reports/ftsAnalysis-destination-kiribati.html)", 
"[Profile Report for destination: Korea, Democratic People's Republic of](../docs/reports/ftsAnalysis-destination-korea-democratic-peoples-republic-of.html)", 
"[Profile Report for destination: Kuwait](../docs/reports/ftsAnalysis-destination-kuwait.html)", 
"[Profile Report for destination: Kyrgyzstan](../docs/reports/ftsAnalysis-destination-kyrgyzstan.html)", 
"[Profile Report for destination: Lao People's Democratic Republic](../docs/reports/ftsAnalysis-destination-lao-peoples-democratic-republic.html)", 
"[Profile Report for destination: Lebanon](../docs/reports/ftsAnalysis-destination-lebanon.html)", 
"[Profile Report for destination: Lesotho](../docs/reports/ftsAnalysis-destination-lesotho.html)", 
"[Profile Report for destination: Liberia](../docs/reports/ftsAnalysis-destination-liberia.html)", 
"[Profile Report for destination: Libya](../docs/reports/ftsAnalysis-destination-libya.html)", 
"[Profile Report for destination: Madagascar](../docs/reports/ftsAnalysis-destination-madagascar.html)", 
"[Profile Report for destination: Malawi](../docs/reports/ftsAnalysis-destination-malawi.html)", 
"[Profile Report for destination: Malaysia](../docs/reports/ftsAnalysis-destination-malaysia.html)", 
"[Profile Report for destination: Maldives](../docs/reports/ftsAnalysis-destination-maldives.html)", 
"[Profile Report for destination: Mali](../docs/reports/ftsAnalysis-destination-mali.html)", 
"[Profile Report for destination: Marshall Islands](../docs/reports/ftsAnalysis-destination-marshall-islands.html)", 
"[Profile Report for destination: Mauritania](../docs/reports/ftsAnalysis-destination-mauritania.html)", 
"[Profile Report for destination: Mauritius](../docs/reports/ftsAnalysis-destination-mauritius.html)", 
"[Profile Report for destination: Mexico](../docs/reports/ftsAnalysis-destination-mexico.html)", 
"[Profile Report for destination: Micronesia, Federated States of](../docs/reports/ftsAnalysis-destination-micronesia-federated-states-of.html)", 
"[Profile Report for destination: Moldova, Republic of](../docs/reports/ftsAnalysis-destination-moldova-republic-of.html)", 
"[Profile Report for destination: Mongolia](../docs/reports/ftsAnalysis-destination-mongolia.html)", 
"[Profile Report for destination: Montenegro](../docs/reports/ftsAnalysis-destination-montenegro.html)", 
"[Profile Report for destination: Morocco](../docs/reports/ftsAnalysis-destination-morocco.html)", 
"[Profile Report for destination: Mozambique](../docs/reports/ftsAnalysis-destination-mozambique.html)", 
"[Profile Report for destination: Myanmar](../docs/reports/ftsAnalysis-destination-myanmar.html)", 
"[Profile Report for destination: Namibia](../docs/reports/ftsAnalysis-destination-namibia.html)", 
"[Profile Report for destination: Nepal](../docs/reports/ftsAnalysis-destination-nepal.html)", 
"[Profile Report for destination: Netherlands Antilles](../docs/reports/ftsAnalysis-destination-netherlands-antilles.html)", 
"[Profile Report for destination: New Zealand](../docs/reports/ftsAnalysis-destination-new-zealand.html)", 
"[Profile Report for destination: Nicaragua](../docs/reports/ftsAnalysis-destination-nicaragua.html)", 
"[Profile Report for destination: Niger](../docs/reports/ftsAnalysis-destination-niger.html)", 
"[Profile Report for destination: Nigeria](../docs/reports/ftsAnalysis-destination-nigeria.html)", 
"[Profile Report for destination: Niue](../docs/reports/ftsAnalysis-destination-niue.html)", 
"[Profile Report for destination: North Macedonia, Republic of](../docs/reports/ftsAnalysis-destination-north-macedonia-republic-of.html)", 
"[Profile Report for destination: Norway](../docs/reports/ftsAnalysis-destination-norway.html)", 
"[Profile Report for destination: Occupied Palestinian Territory](../docs/reports/ftsAnalysis-destination-occupied-palestinian-territory.html)", 
"[Profile Report for destination: Oman](../docs/reports/ftsAnalysis-destination-oman.html)", 
"[Profile Report for destination: Pakistan](../docs/reports/ftsAnalysis-destination-pakistan.html)", 
"[Profile Report for destination: Palau](../docs/reports/ftsAnalysis-destination-palau.html)", 
"[Profile Report for destination: Panama](../docs/reports/ftsAnalysis-destination-panama.html)", 
"[Profile Report for destination: Papua New Guinea](../docs/reports/ftsAnalysis-destination-papua-new-guinea.html)", 
"[Profile Report for destination: Paraguay](../docs/reports/ftsAnalysis-destination-paraguay.html)", 
"[Profile Report for destination: Peru](../docs/reports/ftsAnalysis-destination-peru.html)", 
"[Profile Report for destination: Philippines](../docs/reports/ftsAnalysis-destination-philippines.html)", 
"[Profile Report for destination: Poland](../docs/reports/ftsAnalysis-destination-poland.html)", 
"[Profile Report for destination: Qatar](../docs/reports/ftsAnalysis-destination-qatar.html)", 
"[Profile Report for destination: Region - Asia and the Pacific](../docs/reports/ftsAnalysis-destination-region-asia-and-the-pacific.html)", 
"[Profile Report for destination: Region - Europe](../docs/reports/ftsAnalysis-destination-region-europe.html)", 
"[Profile Report for destination: Region - Latin America and the Caribbean](../docs/reports/ftsAnalysis-destination-region-latin-america-and-the-caribbean.html)", 
"[Profile Report for destination: Region - Middle East and Northern Africa](../docs/reports/ftsAnalysis-destination-region-middle-east-and-northern-africa.html)", 
"[Profile Report for destination: Region - Southern and Eastern Africa](../docs/reports/ftsAnalysis-destination-region-southern-and-eastern-africa.html)", 
"[Profile Report for destination: Region - West and Central Africa](../docs/reports/ftsAnalysis-destination-region-west-and-central-africa.html)", 
"[Profile Report for destination: Romania](../docs/reports/ftsAnalysis-destination-romania.html)", 
"[Profile Report for destination: Russian Federation](../docs/reports/ftsAnalysis-destination-russian-federation.html)", 
"[Profile Report for destination: Rwanda](../docs/reports/ftsAnalysis-destination-rwanda.html)", 
"[Profile Report for destination: Saint Lucia](../docs/reports/ftsAnalysis-destination-saint-lucia.html)", 
"[Profile Report for destination: Saint Vincent and the Grenadines](../docs/reports/ftsAnalysis-destination-saint-vincent-and-the-grenadines.html)", 
"[Profile Report for destination: Samoa](../docs/reports/ftsAnalysis-destination-samoa.html)", 
"[Profile Report for destination: Senegal](../docs/reports/ftsAnalysis-destination-senegal.html)", 
"[Profile Report for destination: Serbia](../docs/reports/ftsAnalysis-destination-serbia.html)", 
"[Profile Report for destination: Seychelles](../docs/reports/ftsAnalysis-destination-seychelles.html)", 
"[Profile Report for destination: Sierra Leone](../docs/reports/ftsAnalysis-destination-sierra-leone.html)", 
"[Profile Report for destination: Singapore](../docs/reports/ftsAnalysis-destination-singapore.html)", 
"[Profile Report for destination: Slovakia](../docs/reports/ftsAnalysis-destination-slovakia.html)", 
"[Profile Report for destination: Solomon Islands](../docs/reports/ftsAnalysis-destination-solomon-islands.html)", 
"[Profile Report for destination: Somalia](../docs/reports/ftsAnalysis-destination-somalia.html)", 
"[Profile Report for destination: South Africa](../docs/reports/ftsAnalysis-destination-south-africa.html)", 
"[Profile Report for destination: South Sudan](../docs/reports/ftsAnalysis-destination-south-sudan.html)", 
"[Profile Report for destination: Spain](../docs/reports/ftsAnalysis-destination-spain.html)", 
"[Profile Report for destination: Sri Lanka](../docs/reports/ftsAnalysis-destination-sri-lanka.html)", 
"[Profile Report for destination: Sudan](../docs/reports/ftsAnalysis-destination-sudan.html)", 
"[Profile Report for destination: Suriname](../docs/reports/ftsAnalysis-destination-suriname.html)", 
"[Profile Report for destination: Switzerland](../docs/reports/ftsAnalysis-destination-switzerland.html)", 
"[Profile Report for destination: Syrian Arab Republic](../docs/reports/ftsAnalysis-destination-syrian-arab-republic.html)", 
"[Profile Report for destination: Tajikistan](../docs/reports/ftsAnalysis-destination-tajikistan.html)", 
"[Profile Report for destination: Tanzania, United Republic of](../docs/reports/ftsAnalysis-destination-tanzania-united-republic-of.html)", 
"[Profile Report for destination: Thailand](../docs/reports/ftsAnalysis-destination-thailand.html)", 
"[Profile Report for destination: Timor-Leste](../docs/reports/ftsAnalysis-destination-timor-leste.html)", 
"[Profile Report for destination: Togo](../docs/reports/ftsAnalysis-destination-togo.html)", 
"[Profile Report for destination: Tonga](../docs/reports/ftsAnalysis-destination-tonga.html)", 
"[Profile Report for destination: Trinidad and Tobago](../docs/reports/ftsAnalysis-destination-trinidad-and-tobago.html)", 
"[Profile Report for destination: Tunisia](../docs/reports/ftsAnalysis-destination-tunisia.html)", 
"[Profile Report for destination: Türkiye](../docs/reports/ftsAnalysis-destination-turkiye.html)", 
"[Profile Report for destination: Turkmenistan](../docs/reports/ftsAnalysis-destination-turkmenistan.html)", 
"[Profile Report for destination: Uganda](../docs/reports/ftsAnalysis-destination-uganda.html)", 
"[Profile Report for destination: Ukraine](../docs/reports/ftsAnalysis-destination-ukraine.html)", 
"[Profile Report for destination: United Arab Emirates](../docs/reports/ftsAnalysis-destination-united-arab-emirates.html)", 
"[Profile Report for destination: United Kingdom](../docs/reports/ftsAnalysis-destination-united-kingdom.html)", 
"[Profile Report for destination: United States](../docs/reports/ftsAnalysis-destination-united-states.html)", 
"[Profile Report for destination: Uruguay](../docs/reports/ftsAnalysis-destination-uruguay.html)", 
"[Profile Report for destination: Uzbekistan](../docs/reports/ftsAnalysis-destination-uzbekistan.html)", 
"[Profile Report for destination: Vanuatu](../docs/reports/ftsAnalysis-destination-vanuatu.html)", 
"[Profile Report for destination: Venezuela, Bolivarian Republic of](../docs/reports/ftsAnalysis-destination-venezuela-bolivarian-republic-of.html)", 
"[Profile Report for destination: Viet Nam](../docs/reports/ftsAnalysis-destination-viet-nam.html)", 
"[Profile Report for destination: Yemen](../docs/reports/ftsAnalysis-destination-yemen.html)", 
"[Profile Report for destination: Zambia](../docs/reports/ftsAnalysis-destination-zambia.html)", 
"[Profile Report for destination: Zimbabwe](../docs/reports/ftsAnalysis-destination-zimbabwe.html)")